@@ -13,80 +12,80 @@
         "tsc": "tsc"
     },
     "devDependencies": {
-        "@types/chai": "^4.1.7",
+        "@types/chai": "^4.2.7",
         "@types/mocha": "^5.2.6",
-        "@types/react-dom": "^16.8.2",
+        "@types/react-dom": "^16.9.4",
         "@types/webpack-dev-middleware": "^2.0.2",
-        "@types/webpack-hot-middleware": "^2.16.4",
+        "@types/webpack-hot-middleware": "^2.25.0",
         "awesome-typescript-loader": "^5.2.1",
         "chai": "^4.2.0",
         "copy-webpack-plugin": "^4.6.0",
-        "cross-env": "^5.2.0",
+        "cross-env": "^5.2.1",
         "css-loader": "^2.1.1",
         "file-loader": "^3.0.1",
-        "fork-ts-checker-webpack-plugin": "^1.0.2",
-        "jsdom": "^15.1.1",
+        "fork-ts-checker-webpack-plugin": "^1.6.0",
+        "jsdom": "^15.2.1",
         "mocha": "^5.2.0",
-        "sass-loader": "^7.1.0",
+        "sass-loader": "^7.3.1",
         "scss-loader": "0.0.1",
         "style-loader": "^0.23.1",
         "ts-loader": "^5.3.3",
         "ts-node": "^7.0.1",
-        "ts-node-dev": "^1.0.0-pre.32",
-        "tslint": "^5.15.0",
+        "ts-node-dev": "^1.0.0-pre.44",
+        "tslint": "^5.20.1",
         "tslint-loader": "^3.5.4",
-        "typescript": "^3.7.2",
-        "webpack": "^4.29.6",
-        "webpack-cli": "^3.2.3",
-        "webpack-dev-middleware": "^3.6.1",
-        "webpack-dev-server": "^3.3.1",
+        "typescript": "^3.7.4",
+        "webpack": "^4.41.5",
+        "webpack-cli": "^3.3.10",
+        "webpack-dev-middleware": "^3.7.2",
+        "webpack-dev-server": "^3.10.1",
         "webpack-hot-middleware": "^2.24.3"
     },
     "dependencies": {
         "@fortawesome/fontawesome-free-solid": "^5.0.13",
-        "@fortawesome/fontawesome-svg-core": "^1.2.15",
-        "@fortawesome/free-brands-svg-icons": "^5.9.0",
-        "@fortawesome/free-regular-svg-icons": "^5.9.0",
-        "@fortawesome/free-solid-svg-icons": "^5.7.2",
-        "@fortawesome/react-fontawesome": "^0.1.4",
+        "@fortawesome/fontawesome-svg-core": "^1.2.26",
+        "@fortawesome/free-brands-svg-icons": "^5.12.0",
+        "@fortawesome/free-regular-svg-icons": "^5.12.0",
+        "@fortawesome/free-solid-svg-icons": "^5.12.0",
+        "@fortawesome/react-fontawesome": "^0.1.8",
         "@hig/flyout": "^1.2.0",
         "@hig/theme-context": "^2.1.3",
-        "@hig/theme-data": "^2.3.3",
+        "@hig/theme-data": "^2.13.0",
         "@trendmicro/react-dropdown": "^1.3.0",
         "@types/adm-zip": "^0.4.32",
         "@types/animejs": "^2.0.2",
         "@types/archiver": "^3.0.0",
         "@types/async": "^2.4.1",
         "@types/bcrypt-nodejs": "0.0.30",
-        "@types/bluebird": "^3.5.25",
-        "@types/body-parser": "^1.17.0",
+        "@types/bluebird": "^3.5.29",
+        "@types/body-parser": "^1.17.1",
         "@types/classnames": "^2.2.8",
-        "@types/color": "^3.0.0",
+        "@types/color": "^3.0.1",
         "@types/connect-flash": "0.0.34",
-        "@types/cookie-parser": "^1.4.1",
+        "@types/cookie-parser": "^1.4.2",
         "@types/cookie-session": "^2.0.36",
         "@types/d3-format": "^1.3.1",
         "@types/dotenv": "^6.1.1",
         "@types/exif": "^0.6.0",
-        "@types/express": "^4.16.1",
+        "@types/express": "^4.17.2",
         "@types/express-flash": "0.0.0",
-        "@types/express-session": "^1.15.12",
+        "@types/express-session": "^1.15.16",
         "@types/express-validator": "^3.0.0",
         "@types/formidable": "^1.0.31",
         "@types/gapi": "0.0.39",
-        "@types/jquery": "^3.3.29",
+        "@types/jquery": "^3.3.31",
         "@types/jquery-awesome-cursor": "^0.3.0",
-        "@types/jsonwebtoken": "^8.3.2",
-        "@types/lodash": "^4.14.121",
+        "@types/jsonwebtoken": "^8.3.5",
+        "@types/lodash": "^4.14.149",
         "@types/mobile-detect": "^1.3.4",
-        "@types/mongodb": "^3.1.22",
-        "@types/mongoose": "^5.5.8",
-        "@types/node": "^10.12.30",
+        "@types/mongodb": "^3.3.14",
+        "@types/mongoose": "^5.5.38",
+        "@types/node": "^10.17.13",
         "@types/nodemailer": "^4.6.6",
         "@types/passport": "^1.0.2",
-        "@types/passport-google-oauth20": "^2.0.2",
+        "@types/passport-google-oauth20": "^2.0.3",
         "@types/passport-local": "^1.0.33",
-        "@types/pdfjs-dist": "^2.0.0",
+        "@types/pdfjs-dist": "^2.1.3",
         "@types/prosemirror-commands": "^1.0.1",
         "@types/prosemirror-history": "^1.0.1",
         "@types/prosemirror-inputrules": "^1.0.2",
@@ -165,70 +165,70 @@
         "jsx-to-string": "^1.4.0",
         "kill-port": "^1.6.0",
         "lodash": "^4.17.15",
-        "mobile-detect": "^1.4.3",
-        "mobx": "^5.15.0",
+        "mobile-detect": "^1.4.4",
+        "mobx": "^5.15.1",
         "mobx-react": "^5.3.5",
         "mobx-react-devtools": "^6.1.1",
         "mobx-utils": "^5.5.2",
-        "mongodb": "^3.1.13",
-        "mongoose": "^5.6.4",
-        "node-sass": "^4.12.0",
+        "mongodb": "^3.4.1",
+        "mongoose": "^5.8.4",
+        "node-sass": "^4.13.0",
         "nodemailer": "^5.1.1",
-        "nodemon": "^1.18.10",
+        "nodemon": "^1.19.4",
         "normalize.css": "^8.0.1",
-        "npm": "^6.13.4",
+        "npm": "^6.13.6",
         "p-limit": "^2.2.0",
         "passport": "^0.4.0",
         "passport-google-oauth20": "^2.0.0",
         "passport-local": "^1.0.0",
         "pdf-parse": "^1.1.1",
-        "pdfjs-dist": "^2.0.943",
+        "pdfjs-dist": "^2.2.228",
         "probe-image-size": "^4.0.0",
-        "prosemirror-commands": "^1.0.8",
-        "prosemirror-example-setup": "^1.0.1",
+        "prosemirror-commands": "^1.1.3",
+        "prosemirror-example-setup": "^1.1.2",
         "prosemirror-find-replace": "^0.9.0",
-        "prosemirror-history": "^1.0.4",
-        "prosemirror-keymap": "^1.0.1",
-        "prosemirror-model": "^1.7.2",
-        "prosemirror-schema-basic": "^1.0.1",
-        "prosemirror-schema-list": "^1.0.3",
-        "prosemirror-state": "^1.2.4",
-        "prosemirror-transform": "^1.1.4",
-        "prosemirror-view": "^1.10.3",
+        "prosemirror-history": "^1.1.3",
+        "prosemirror-keymap": "^1.1.3",
+        "prosemirror-model": "^1.8.2",
+        "prosemirror-schema-basic": "^1.1.2",
+        "prosemirror-schema-list": "^1.1.2",
+        "prosemirror-state": "^1.3.2",
+        "prosemirror-transform": "^1.2.3",
+        "prosemirror-view": "^1.13.7",
         "pug": "^2.0.3",
-        "query-string": "^6.8.1",
+        "query-string": "^6.9.0",
         "raw-loader": "^1.0.0",
         "rc-switch": "^1.9.0",
-        "react": "^16.8.4",
-        "react-anime": "^2.2.0",
+        "react": "^16.12.0",
+        "react-anime": "^2.3.2",
         "react-autosuggest": "^9.4.3",
-        "react-bootstrap": "^1.0.0-beta.5",
+        "react-bootstrap": "^1.0.0-beta.16",
         "react-bootstrap-dropdown-menu": "^1.1.15",
         "react-color": "^2.17.0",
         "react-dimensions": "^1.3.1",
-        "react-dom": "^16.8.4",
+        "react-dom": "^16.12.0",
         "react-golden-layout": "^1.0.6",
         "react-image-lightbox-with-rotate": "^5.1.1",
-        "react-jsx-parser": "^1.15.0",
+        "react-jsx-parser": "^1.21.0",
         "react-measure": "^2.2.4",
         "react-mosaic": "0.0.20",
         "react-simple-dropdown": "^3.2.3",
-        "react-split-pane": "^0.1.85",
-        "react-table": "^6.10.3",
+        "react-split-pane": "^0.1.89",
+        "react-table": "^6.11.5",
         "readline": "^1.3.0",
         "request": "^2.88.0",
-        "request-promise": "^4.2.4",
+        "request-promise": "^4.2.5",
         "rimraf": "^3.0.0",
-        "serializr": "^1.5.1",
-        "sharp": "^0.22.1",
+        "serializr": "^1.5.4",
+        "sharp": "^0.23.4",
         "shelljs": "^0.8.3",
-        "socket.io": "^2.2.0",
-        "socket.io-client": "^2.2.0",
+        "socket.io": "^2.3.0",
+        "socket.io-client": "^2.3.0",
         "solr-node": "^1.2.1",
         "standard-http-error": "^2.0.1",
-        "typescript-collections": "^1.3.2",
+        "typescript-collections": "^1.3.3",
         "url-loader": "^1.1.2",
-        "uuid": "^3.3.2",
+        "uuid": "^3.3.3",
         "wikijs": "^6.0.1",
         "words-to-numbers": "^1.5.1",
         "xoauth2": "^1.2.0",
