@@ -6,6 +6,7 @@ import * as rimraf from "rimraf";
 import { yellow, Color } from 'colors';
 import * as nodemailer from "nodemailer";
 import { MailOptions } from "nodemailer/lib/json-transport";
+import Mail = require('nodemailer/lib/mailer');
 
 const projectRoot = path.resolve(__dirname, "../../");
 export function pathFromRoot(relative?: string) {
@@ -118,20 +119,42 @@ export namespace Email {
         }
     });
 
-    export async function dispatchAll(recipients: string[], subject: string, content: string) {
-        return Promise.all(recipients.map((recipient: string) => Email.dispatch(recipient, subject, content)));
+    export interface DispatchOptions<T extends string | string[]> {
+        to: T;
+        subject: string;
+        content: string;
+        attachments?: Mail.Attachment | Mail.Attachment[];
     }
 
-    export async function dispatch(recipient: string, subject: string, content: string): Promise<boolean> {
+    export interface DispatchFailure {
+        recipient: string;
+        error: Error;
+    }
+
+    export async function dispatchAll({ to, subject, content, attachments }: DispatchOptions<string[]>) {
+        const failures: DispatchFailure[] = [];
+        await Promise.all(to.map(async recipient => {
+            let error: Error | null;
+            const resolved = attachments ? "length" in attachments ? attachments : [attachments] : undefined;
+            if ((error = await Email.dispatch({ to: recipient, subject, content, attachments: resolved })) !== null) {
+                failures.push({
+                    recipient,
+                    error
+                });
+            }
+        }));
+        return failures.length ? failures : undefined;
+    }
+
+    export async function dispatch({ to, subject, content, attachments }: DispatchOptions<string>): Promise<Error | null> {
         const mailOptions = {
-            to: recipient,
+            to,
             from: 'brownptcdash@gmail.com',
             subject,
-            text: `Hello ${recipient.split("@")[0]},\n\n${content}`
+            text: `Hello ${to.split("@")[0]},\n\n${content}`,
+            attachments
         } as MailOptions;
-        return new Promise<boolean>(resolve => {
-            smtpTransport.sendMail(mailOptions, (dispatchError: Error | null) => resolve(dispatchError === null));
-        });
+        return new Promise<Error | null>(resolve => smtpTransport.sendMail(mailOptions, resolve));
     }
 
 }
